My attempt at the unity game engine code
Functions to; Create a game object (sphere). Input the data, to define the sphere, from a file. Calculate the velocity per frame to use for the movement of the spheres. Update the position of the object by adding the VPF and the original position. Update the scale of the sphere based on its mass.